requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
langchain>=0.0.343
pydantic>=1.10.7
python-dotenv>=1.0.0